hierkeys: extract rotation set utilities to rotations.go
Move MasterRotationsForBase and DecomposeRotation out of hierkeys.go
into a dedicated rotations.go. They're rotation-set arithmetic, not
shared "everything" utilities, and they share no helpers with the
remaining contents of hierkeys.go.

Pure file move; no behavior or signature change.
